[update]getMealRecordsの戻り値の型を修正

+++ app/Services/Meal/getMealService.php
@@ -37,9 +37,9 @@ class getMealService implements getMealServiceInterface
      * 指定された栄養素の合計値を計算するメソッド
      * @param Collection $meals
      * @param string $nutrient
-     * @return float
+     * @return int
      */
-    public function calculateNutrientSum(Collection $meals, string $nutrient): float
+    public function calculateNutrientSum(Collection $meals, string $nutrient): int
     {
         return $meals->sum(function ($meal) use ($nutrient) {
             return $meal->foods->sum(function ($food) use ($nutrient) {
